requirements.txt
fastapi
uvicorn
python-dotenv
gunicorn
requests
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
gspread